Support passing show_diff to cron::d
At its core, cron::d creates a file resource. By default, this resource
has show_diff=>true, which causes Puppet to display a diff when its
contents change. It is possible for this file to contain sensitive
information that is not desirable to have printed out to the terminal
when Puppet runs. This change adds support for passing the show_diff
argument from cron::d to cron::file to the file resource.